requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==1.5.3
requests==2.31.0
kafka-python==2.0.2
python-dotenv==1.0.0
gunicorn==21.2.0
snowflake-connector-python==3.0.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
